+++ app/src/pages/ErrorLoginPage/ErrorLoginPage.tsx
@@ -4,7 +4,7 @@
 import { Container as MuiContainer, CssBaseline, Stack, styled, ThemeProvider, useTheme } from '@mui/material';
 import { WarningIcon } from '@opentalk/common';
 import { EventTypeError, useAuth } from '@opentalk/react-redux-appauth';
-import React, { useEffect, useMemo } from 'react';
+import React, { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { ReactComponent as Logo } from '../../assets/images/logo.svg';
@@ -31,12 +31,11 @@ const Container = styled(MuiContainer)(({ theme }) => ({
 
 interface ErrorLoginProps {
   error: Error;
-  redirectUrl: string;
 }
 
 const ERROR_MESSAGE_DISPLAY_BEFORE_SIGNOUT = 2500;
 
-const ErrorLoginPage = ({ error, redirectUrl }: ErrorLoginProps) => {
+const ErrorLoginPage = ({ error }: ErrorLoginProps) => {
   const { t } = useTranslation();
   const theme = useTheme();
   const dispatch = useAppDispatch();
@@ -46,29 +45,16 @@ const ErrorLoginPage = ({ error, redirectUrl }: ErrorLoginProps) => {
     dispatch(hangUp());
   }, [dispatch]);
 
-  const errorDetails = useMemo(() => {
-    switch (error.name) {
-      case EventTypeError.SessionExpired:
-        setTimeout(() => {
-          sessionStorage.setItem('redirect-uri', redirectUrl);
-          signOut();
-        }, ERROR_MESSAGE_DISPLAY_BEFORE_SIGNOUT);
-        return {
-          name: t('error-session-expired'),
-          message: t('error-session-expired-message'),
-        };
-      case 'error_oidc_configuration':
-        return {
-          name: t('error-oidc-configuration'),
-          message: t('error-oidc-configuration-message'),
-        };
-      default:
-        console.error('unknown login error:', error);
-        return {
-          name: t('error-general'),
-        };
+  useEffect(() => {
+    if (error.name === EventTypeError.SessionExpired) {
+      setTimeout(() => {
+        (async () => {
+          sessionStorage.setItem('redirect-uri', window.location.pathname);
+          await signOut();
+        })();
+      }, ERROR_MESSAGE_DISPLAY_BEFORE_SIGNOUT);
     }
-  }, [error, t]);
+  }, [error]);
 
   return (
     <ThemeProvider theme={createOpenTalkTheme()}>
@@ -77,7 +63,7 @@ const ErrorLoginPage = ({ error, redirectUrl }: ErrorLoginProps) => {
         <Container>
           <Stack spacing={8} alignItems="center">
             <WarningIcon fill={theme.palette.common.white} sx={{ width: '6rem', height: '6rem' }} />
-            {errorDetails && <Error title={errorDetails.name} description={errorDetails.message} />}
+            {error && <Error title={t(`${error.name}`)} description={t(`${error.message}`)} />}
             <StyledLogoIcon />
           </Stack>
         </Container>