+++ app/src/store/index.ts
@@ -17,29 +17,22 @@ import configReducer from './slices/configSlice';
 import connectionStatsReducer from './slices/connectionStatsSlice';
 import eventReducer from './slices/eventSlice';
 import mediaReducer from './slices/mediaSlice';
-import subscribersReducer, { mediaSubscriberMiddleware } from './slices/mediaSubscriberSlice';
+import subscribersReducer from './slices/mediaSubscriberSlice';
 import moderationReducer from './slices/moderationSlice';
 import participantsReducer from './slices/participantsSlice';
 import pollReducer from './slices/pollSlice';
 import protocolReducer from './slices/protocolSlice';
 import recordingReducer from './slices/recordingSlice';
-import roomReducer, { roomMiddleware } from './slices/roomSlice';
+import roomReducer from './slices/roomSlice';
 import sharedFolderReducer from './slices/sharedFolderSlice';
 import slotReducer from './slices/slotSlice';
 import speedMeterReducer from './slices/speedMeterSlice';
-import timerReducer, { timerMiddleware } from './slices/timerSlice';
+import timerReducer from './slices/timerSlice';
 import uiReducer from './slices/uiSlice';
 import userReducer from './slices/userSlice';
 import whiteboardReducer from './slices/whiteboardSlice';
 
-const middleware: Array<Middleware> = [
-  apiMiddleware,
-  restApi.middleware,
-  rtkQueryErrorLoggerMiddlware,
-  mediaSubscriberMiddleware.middleware,
-  timerMiddleware.middleware,
-  roomMiddleware.middleware,
-];
+const middleware: Array<Middleware> = [apiMiddleware, restApi.middleware, rtkQueryErrorLoggerMiddlware];
 
 const logger = () => (next: Dispatch) => (action: AnyAction) => {
   if (!action.type.startsWith('stats/statsUpdated')) {